Apply suggestion from @phonofidelic

+++ LMS.Blazor/LMS.Blazor.Client/Pages/CourseDetailsPage/CourseDetailsPage.razor.cs
@@ -92,6 +92,7 @@ public partial class CourseDetailsPage
     }
     private async Task DeleteThis()
     {
+if (CourseDetails != null)
         await ApiService.DeleteAsync($"api/courses/{CourseDetails?.Id}");
         Navigation.NavigateTo("/dashboard");
     }
